Removed some old unused code.
This part of the code was used for some features that are no longer
part of this tool.
Also added a tcl command, to see memory stats (memory usage) after
you complete filtering_process.
Also added filepath file to git (nothing special, just to avoid
an message on the first time of running as a new user).
+ some micro changes