Added global variable as argument on memtrace
Global variable completepath (source directory path) has been added as
argument on memory_trace executable. For unknown reason memory_trace
could not detect the global variable even is before and after the call
it was recognized from all the other functions. This happened on
specific examples, a minority. I could not find the problem, so i just
send a global variable as argument and copy it back to itself.