Connected to optimized gleipnir, edit bug for big apps
In this commit, the tool has been connected to optimized gleipnir which
is stored in github.com/pavlosaim/memory-tracing repository. You can go
there for more information.

Big applications could not fit into some stack variables, changed them
to mallocs. There are more things to be changed to malloc, which will be
done later. There are many hard coded stack variables, not optimal, and
will soon be changed on following commits.

I did a mini research and found that _zzq_args that appear on
visualizations are arguments on valgrind function calls. You can see
them on statistics and infos as for some may be useful, but decided not
to visualize them, as they were not visualizable and destroyed the whole
plot.

Updated README.md with new gleipnir repo info.